requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
tensorflow==2.13.0
numpy>=1.22.0
pillow>=9.0.0
python-multipart>=0.0.5
opencv-python>=4.5.0
pydantic<2.0.0
typing-extensions<4.7.0
